src/Notifier/requirements.txt
requests==2.25.0
boto3==1.16.16
logging-formatter-anticrlf==1.2
